Rename components and add page /work and /our-team
